+++ FantaF1DataAccessDB/IscrizioniScuderieCampionato.cs
@@ -17,7 +17,7 @@ namespace FantaF1DataAccessDB
         public int Id { get; set; }
         public int ScuderiaId { get; set; }
         public int CampionatoId { get; set; }
-        public int Punteggio { get; set; }
+        public decimal Punteggio { get; set; }
     
         public virtual CampionatiMondiali CampionatiMondiali { get; set; }
         public virtual Scuderie Scuderie { get; set; }
